requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.3.0
django-filter==24.2
whitenoise==6.6.0
gunicorn==21.2.0